refactor(certificateutil): 4. split parsing into certificate.go (verbatim move)
util.go mixes two concerns: pure certificate parsing and macOS `security`
keychain access. Separate them so each file has one responsibility and the
keychain half can later be made injectable in isolation. Pure relocation, no
logic changes; cleanups follow next.
